move types to primitives
