+++ crates/net/eth-wire/src/types/transactions.rs
@@ -5,8 +5,9 @@ use bytes::{Buf, Bytes};
 use reth_codecs::derive_arbitrary;
 use reth_primitives::{
     kzg::{self, KzgCommitment, KzgProof, KzgSettings},
-    kzg_to_versioned_hash, BlobTransactionSidecar, Signature, Transaction, TransactionSigned,
-    TransactionSignedNoHash, TxEip4844, TxType, EIP4844_TX_TYPE_ID, H256,
+    kzg_to_versioned_hash, BlobTransactionSidecar, PooledTransactionResponse, Signature,
+    Transaction, TransactionSigned, TransactionSignedNoHash, TxEip4844, TxType, EIP4844_TX_TYPE_ID,
+    H256,
 };
 use reth_rlp::{
     Decodable, DecodeError, Encodable, Header, RlpDecodableWrapper, RlpEncodableWrapper,
@@ -68,534 +69,6 @@ impl From<Vec<TransactionSigned>> for PooledTransactions {
     }
 }
 
-/// A response to [`GetPooledTransactions`]. This can include either a blob transaction, or a
-/// non-4844 signed transaction.
-// TODO: redo arbitrary for this encoding - the previous encoding was incorrect. then this can be
-// replaced with `TransactionSigned` in `PooledTransactions`
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum PooledTransactionResponse {
-    /// A blob transaction, which includes the transaction, blob data, commitments, and proofs.
-    BlobTransaction(BlobTransaction),
-    /// A non-4844 signed transaction.
-    Transaction(TransactionSigned),
-}
-
-impl PooledTransactionResponse {
-    /// Decodes the "raw" format of transaction (e.g. `eth_sendRawTransaction`).
-    ///
-    /// The raw transaction is either a legacy transaction or EIP-2718 typed transaction
-    /// For legacy transactions, the format is encoded as: `rlp(tx)`
-    /// For EIP-2718 typed transaction, the format is encoded as the type of the transaction
-    /// followed by the rlp of the transaction: `type` + `rlp(tx)`
-    ///
-    /// For encoded EIP-4844 transactions, the blob sidecar _must_ be included.
-    pub fn decode_enveloped(tx: Bytes) -> Result<Self, DecodeError> {
-        let mut data = tx.as_ref();
-
-        if data.is_empty() {
-            return Err(DecodeError::InputTooShort)
-        }
-
-        // Check if the tx is a list - tx types are less than EMPTY_LIST_CODE (0xc0)
-        if data[0] >= EMPTY_LIST_CODE {
-            // decode as legacy transaction
-            Ok(Self::Transaction(TransactionSigned::decode_rlp_legacy_transaction(&mut data)?))
-        } else {
-            // decode the type byte, only decode BlobTransaction if it is a 4844 transaction
-            let tx_type = *data.first().ok_or(DecodeError::InputTooShort)?;
-
-            if tx_type == EIP4844_TX_TYPE_ID {
-                // Recall that the blob transaction response `TranactionPayload` is encoded like
-                // this: `rlp([tx_payload_body, blobs, commitments, proofs])`
-                //
-                // Note that `tx_payload_body` is a list:
-                // `[chain_id, nonce, max_priority_fee_per_gas, ..., y_parity, r, s]`
-                //
-                // This makes the full encoding:
-                // `tx_type (0x03) || rlp([[chain_id, nonce, ...], blobs, commitments, proofs])`
-                //
-                // First, we advance the buffer past the type byte
-                data.advance(1);
-
-                // Now, we decode the inner blob transaction:
-                // `rlp([[chain_id, nonce, ...], blobs, commitments, proofs])`
-                let blob_tx = BlobTransaction::decode_inner(&mut data)?;
-                Ok(PooledTransactionResponse::BlobTransaction(blob_tx))
-            } else {
-                // DO NOT advance the buffer for the type, since we want the enveloped decoding to
-                // decode it again and advance the buffer on its own.
-                let typed_tx = TransactionSigned::decode_enveloped_typed_transaction(&mut data)?;
-                Ok(PooledTransactionResponse::Transaction(typed_tx))
-            }
-        }
-    }
-
-    /// Returns the inner [TransactionSigned].
-    pub fn into_transaction(self) -> TransactionSigned {
-        match self {
-            Self::Transaction(tx) => tx,
-            Self::BlobTransaction(blob_tx) => blob_tx.transaction,
-        }
-    }
-}
-
-impl Encodable for PooledTransactionResponse {
-    /// Encodes an enveloped post EIP-4844 [PooledTransactionResponse].
-    fn encode(&self, out: &mut dyn bytes::BufMut) {
-        match self {
-            Self::Transaction(tx) => tx.encode(out),
-            Self::BlobTransaction(blob_tx) => {
-                // The inner encoding is used with `with_header` set to true, making the final
-                // encoding:
-                // `rlp(tx_type || rlp([transaction_payload_body, blobs, commitments, proofs]))`
-                blob_tx.encode_with_type_inner(out, true);
-            }
-        }
-    }
-
-    fn length(&self) -> usize {
-        match self {
-            Self::Transaction(tx) => tx.length(),
-            Self::BlobTransaction(blob_tx) => {
-                // the encoding uses a header, so we set `with_header` to true
-                blob_tx.payload_len_with_type(true)
-            }
-        }
-    }
-}
-
-impl Decodable for PooledTransactionResponse {
-    /// Decodes an enveloped post EIP-4844 [PooledTransactionResponse].
-    ///
-    /// CAUTION: this expects that `buf` is `[id, rlp(tx)]`
-    fn decode(buf: &mut &[u8]) -> Result<Self, DecodeError> {
-        // From the EIP-4844 spec:
-        // Blob transactions have two network representations. During transaction gossip responses
-        // (`PooledTransactions`), the EIP-2718 `TransactionPayload` of the blob transaction is
-        // wrapped to become:
-        //
-        // `rlp([tx_payload_body, blobs, commitments, proofs])`
-        //
-        // This means the full wire encoding is:
-        // `rlp(tx_type || rlp([transaction_payload_body, blobs, commitments, proofs]))`
-        //
-        // First, we check whether or not the transaction is a legacy transaction.
-        if buf.is_empty() {
-            return Err(DecodeError::InputTooShort)
-        }
-
-        // keep this around for buffer advancement post-legacy decoding
-        let mut original_encoding = *buf;
-
-        // If the header is a list header, it is a legacy transaction. Otherwise, it is a typed
-        // transaction
-        let header = Header::decode(buf)?;
-
-        // Check if the tx is a list
-        if header.list {
-            // decode as legacy transaction
-            let legacy_tx =
-                TransactionSigned::decode_rlp_legacy_transaction(&mut original_encoding)?;
-
-            // advance the buffer based on how far `decode_rlp_legacy_transaction` advanced the
-            // buffer
-            *buf = original_encoding;
-
-            Ok(PooledTransactionResponse::Transaction(legacy_tx))
-        } else {
-            // decode the type byte, only decode BlobTransaction if it is a 4844 transaction
-            let tx_type = *buf.first().ok_or(DecodeError::InputTooShort)?;
-
-            if tx_type == EIP4844_TX_TYPE_ID {
-                // Recall that the blob transaction response `TranactionPayload` is encoded like
-                // this: `rlp([tx_payload_body, blobs, commitments, proofs])`
-                //
-                // Note that `tx_payload_body` is a list:
-                // `[chain_id, nonce, max_priority_fee_per_gas, ..., y_parity, r, s]`
-                //
-                // This makes the full encoding:
-                // `tx_type (0x03) || rlp([[chain_id, nonce, ...], blobs, commitments, proofs])`
-                //
-                // First, we advance the buffer past the type byte
-                buf.advance(1);
-
-                // Now, we decode the inner blob transaction:
-                // `rlp([[chain_id, nonce, ...], blobs, commitments, proofs])`
-                let blob_tx = BlobTransaction::decode_inner(buf)?;
-                Ok(PooledTransactionResponse::BlobTransaction(blob_tx))
-            } else {
-                // DO NOT advance the buffer for the type, since we want the enveloped decoding to
-                // decode it again and advance the buffer on its own.
-                let typed_tx = TransactionSigned::decode_enveloped_typed_transaction(buf)?;
-                Ok(PooledTransactionResponse::Transaction(typed_tx))
-            }
-        }
-    }
-}
-
-impl From<TransactionSigned> for PooledTransactionResponse {
-    /// Converts from a [TransactionSigned] to a [PooledTransactionResponse].
-    ///
-    /// NOTE: This will always return a [PooledTransactionResponse::Transaction] variant.
-    fn from(tx: TransactionSigned) -> Self {
-        Self::Transaction(tx)
-    }
-}
-
-/// An error that can occur when validating a [BlobTransaction].
-#[derive(Debug)]
-pub enum BlobTransactionValidationError {
-    /// An error returned by the [kzg] library
-    KZGError(kzg::Error),
-    /// The inner transaction is not a blob transaction
-    NotBlobTransaction(TxType),
-}
-
-impl From<kzg::Error> for BlobTransactionValidationError {
-    fn from(value: kzg::Error) -> Self {
-        Self::KZGError(value)
-    }
-}
-
-/// A response to [`GetPooledTransactions`] that includes blob data, their commitments, and their
-/// corresponding proofs.
-///
-/// This is defined in [EIP-4844](https://eips.ethereum.org/EIPS/eip-4844#networking) as an element
-/// of a [PooledTransactions] response.
-///
-/// NOTE: This contains a [TransactionSigned], which could be a non-4844 transaction type, even
-/// though that would not make sense. This type is meant to be constructed using decoding methods,
-/// which should always construct the [TransactionSigned] with an EIP-4844 transaction.
-#[derive(Clone, Debug, PartialEq, Eq, Default)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct BlobTransaction {
-    /// The transaction payload.
-    pub transaction: TransactionSigned,
-    /// The transaction's blob sidecar.
-    pub sidecar: BlobTransactionSidecar,
-}
-
-impl BlobTransaction {
-    /// Verifies that the transaction's blob data, commitments, and proofs are all valid.
-    ///
-    /// Takes as input the [KzgSettings], which should contain the the parameters derived from the
-    /// KZG trusted setup.
-    ///
-    /// This ensures that the blob transaction payload has the same number of blob data elements,
-    /// commitments, and proofs. Each blob data element is verified against its commitment and
-    /// proof.
-    ///
-    /// Returns `false` if any blob KZG proof in the response fails to verify, or if the versioned
-    /// hashes in the transaction do not match the actual commitment versioned hashes.
-    pub fn validate(
-        &self,
-        proof_settings: &KzgSettings,
-    ) -> Result<bool, BlobTransactionValidationError> {
-        let inner_tx = match &self.transaction.transaction {
-            Transaction::Eip4844(blob_tx) => blob_tx,
-            non_blob_tx => {
-                return Err(BlobTransactionValidationError::NotBlobTransaction(
-                    non_blob_tx.tx_type(),
-                ))
-            }
-        };
-
-        // Ensure the versioned hashes and commitments have the same length
-        if inner_tx.blob_versioned_hashes.len() != self.sidecar.commitments.len() {
-            return Err(kzg::Error::MismatchLength(format!(
-                "There are {} versioned commitment hashes and {} commitments",
-                inner_tx.blob_versioned_hashes.len(),
-                self.sidecar.commitments.len()
-            ))
-            .into())
-        }
-
-        // zip and iterate, calculating versioned hashes
-        for (versioned_hash, commitment) in
-            inner_tx.blob_versioned_hashes.iter().zip(self.sidecar.commitments.iter())
-        {
-            // convert to KzgCommitment
-            let commitment = KzgCommitment::from(*commitment.deref());
-
-            // Calculate the versioned hash
-            //
-            // TODO: should this method distinguish the type of validation failure? For example
-            // whether a certain versioned hash does not match, or whether the blob proof
-            // validation failed?
-            let calculated_versioned_hash = kzg_to_versioned_hash(commitment);
-            if *versioned_hash != calculated_versioned_hash {
-                return Ok(false)
-            }
-        }
-
-        // Verify as a batch
-        KzgProof::verify_blob_kzg_proof_batch(
-            self.sidecar.blobs.as_slice(),
-            self.sidecar.commitments.as_slice(),
-            self.sidecar.proofs.as_slice(),
-            proof_settings,
-        )
-        .map_err(Into::into)
-    }
-
-    /// Splits the [BlobTransaction] into its [TransactionSigned] and [BlobTransactionSidecar]
-    /// components.
-    pub fn into_parts(self) -> (TransactionSigned, BlobTransactionSidecar) {
-        (self.transaction, self.sidecar)
-    }
-
-    /// Encodes the [BlobTransaction] fields as RLP, with a tx type. If `with_header` is `false`,
-    /// the following will be encoded:
-    /// `tx_type (0x03) || rlp([transaction_payload_body, blobs, commitments, proofs])`
-    ///
-    /// If `with_header` is `true`, the following will be encoded:
-    /// `rlp(tx_type (0x03) || rlp([transaction_payload_body, blobs, commitments, proofs]))`
-    ///
-    /// NOTE: The header will be a byte string header, not a list header.
-    pub fn encode_with_type_inner(&self, out: &mut dyn bytes::BufMut, with_header: bool) {
-        // Calculate the length of:
-        // `tx_type || rlp([transaction_payload_body, blobs, commitments, proofs])`
-        //
-        // to construct and encode the string header
-        if with_header {
-            Header {
-                list: false,
-                // add one for the tx type
-                payload_length: 1 + self.payload_len(),
-            }
-            .encode(out);
-        }
-
-        out.put_u8(EIP4844_TX_TYPE_ID);
-
-        // Now we encode the inner blob transaction:
-        self.encode_inner(out);
-    }
-
-    /// Encodes the [BlobTransaction] fields as RLP, with the following format:
-    /// `rlp([transaction_payload_body, blobs, commitments, proofs])`
-    ///
-    /// where `transaction_payload_body` is a list:
-    /// `[chain_id, nonce, max_priority_fee_per_gas, ..., y_parity, r, s]`
-    ///
-    /// Note: this should be used only when implementing other RLP encoding methods, and does not
-    /// represent the full RLP encoding of the blob transaction.
-    pub fn encode_inner(&self, out: &mut dyn bytes::BufMut) {
-        // First we construct both required list headers.
-        //
-        // The `transaction_payload_body` length is the length of the fields, plus the length of
-        // its list header.
-        let tx_header = Header {
-            list: true,
-            payload_length: self.transaction.fields_len() +
-                self.transaction.signature.payload_len(),
-        };
-
-        let tx_length = tx_header.length() + tx_header.payload_length;
-
-        // The payload length is the length of the `tranascation_payload_body` list, plus the
-        // length of the blobs, commitments, and proofs.
-        let payload_length = tx_length + self.sidecar.fields_len();
-
-        // First we use the payload len to construct the first list header
-        let blob_tx_header = Header { list: true, payload_length };
-
-        // Encode the blob tx header first
-        blob_tx_header.encode(out);
-
-        // Encode the inner tx list header, then its fields
-        tx_header.encode(out);
-        self.transaction.encode_fields(out);
-
-        // Encode the blobs, commitments, and proofs
-        self.sidecar.encode_inner(out);
-    }
-
-    /// Ouputs the length of the RLP encoding of the blob transaction, including the tx type byte,
-    /// optionally including the length of a wrapping string header. If `with_header` is `false`,
-    /// the length of the following will be calculated:
-    /// `tx_type (0x03) || rlp([transaction_payload_body, blobs, commitments, proofs])`
-    ///
-    /// If `with_header` is `true`, the length of the following will be calculated:
-    /// `rlp(tx_type (0x03) || rlp([transaction_payload_body, blobs, commitments, proofs]))`
-    pub fn payload_len_with_type(&self, with_header: bool) -> usize {
-        if with_header {
-            // Construct a header and use that to calculate the total length
-            let wrapped_header = Header {
-                list: false,
-                // add one for the tx type byte
-                payload_length: 1 + self.payload_len(),
-            };
-
-            // The total length is now the length of the header plus the length of the payload
-            // (which includes the tx type byte)
-            wrapped_header.length() + wrapped_header.payload_length
-        } else {
-            // Just add the length of the tx type to the payload length
-            1 + self.payload_len()
-        }
-    }
-
-    /// Outputs the length of the RLP encoding of the blob transaction with the following format:
-    /// `rlp([transaction_payload_body, blobs, commitments, proofs])`
-    ///
-    /// where `transaction_payload_body` is a list:
-    /// `[chain_id, nonce, max_priority_fee_per_gas, ..., y_parity, r, s]`
-    ///
-    /// Note: this should be used only when implementing other RLP encoding length methods, and
-    /// does not represent the full RLP encoding of the blob transaction.
-    pub fn payload_len(&self) -> usize {
-        // The `transaction_payload_body` length is the length of the fields, plus the length of
-        // its list header.
-        let tx_header = Header {
-            list: true,
-            payload_length: self.transaction.fields_len() +
-                self.transaction.signature.payload_len(),
-        };
-
-        let tx_length = tx_header.length() + tx_header.payload_length;
-
-        // The payload length is the length of the `tranascation_payload_body` list, plus the
-        // length of the blobs, commitments, and proofs.
-        tx_length + self.sidecar.fields_len()
-    }
-
-    /// Decodes a [BlobTransaction] from RLP. This expects the encoding to be:
-    /// `rlp([transaction_payload_body, blobs, commitments, proofs])`
-    ///
-    /// where `transaction_payload_body` is a list:
-    /// `[chain_id, nonce, max_priority_fee_per_gas, ..., y_parity, r, s]`
-    ///
-    /// Note: this should be used only when implementing other RLP decoding methods, and does not
-    /// represent the full RLP decoding of the [PooledTransactionResponse] type.
-    pub fn decode_inner(data: &mut &[u8]) -> Result<Self, DecodeError> {
-        // decode the _first_ list header for the rest of the transaction
-        let header = Header::decode(data)?;
-        if !header.list {
-            return Err(DecodeError::Custom("PooledTransactions blob tx must be encoded as a list"))
-        }
-
-        // Now we need to decode the inner 4844 transaction and its signature:
-        //
-        // `[chain_id, nonce, max_priority_fee_per_gas, ..., y_parity, r, s]`
-        let header = Header::decode(data)?;
-        if !header.list {
-            return Err(DecodeError::Custom(
-                "PooledTransactions inner blob tx must be encoded as a list",
-            ))
-        }
-
-        // inner transaction
-        let transaction = Transaction::Eip4844(TxEip4844::decode_inner(data)?);
-
-        // signature
-        let signature = Signature::decode(data)?;
-
-        // construct the tx now that we've decoded the fields in order
-        let tx_no_hash = TransactionSignedNoHash { transaction, signature };
-
-        // All that's left are the blobs, commitments, and proofs
-        let sidecar = BlobTransactionSidecar::decode_inner(data)?;
-
-        // # Calculating the hash
-        //
-        // The full encoding of the [PooledTransaction] response is:
-        // `tx_type (0x03) || rlp([tx_payload_body, blobs, commitments, proofs])`
-        //
-        // The transaction hash however, is:
-        // `keccak256(tx_type (0x03) || rlp(tx_payload_body))`
-        //
-        // Note that this is `tx_payload_body`, not `[tx_payload_body]`, which would be
-        // `[[chain_id, nonce, max_priority_fee_per_gas, ...]]`, i.e. a list within a list.
-        //
-        // Because the pooled transaction encoding is different than the hash encoding for
-        // EIP-4844 transactions, we do not use the original buffer to calculate the hash.
-        //
-        // Instead, we use [TransactionSignedNoHash] which will encode the transaction internally.
-        let signed_tx = tx_no_hash.with_hash();
-
-        Ok(Self { transaction: signed_tx, sidecar })
-    }
-}
-
-#[cfg(any(test, feature = "arbitrary"))]
-impl<'a> arbitrary::Arbitrary<'a> for BlobTransaction {
-    fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
-        let mut arr = [0u8; BYTES_PER_BLOB];
-        let blobs: Vec<Blob> = (0..u.int_in_range(1..=16)?)
-            .map(|_| {
-                arr = arbitrary::Arbitrary::arbitrary(u).unwrap();
-
-                // Ensure that the blob is canonical by ensuring that
-                // each field element contained in the blob is < BLS_MODULUS
-                for i in 0..(FIELD_ELEMENTS_PER_BLOB as usize) {
-                    arr[i * BYTES_PER_FIELD_ELEMENT] = 0;
-                }
-                Blob::from(arr)
-            })
-            .collect();
-
-        Ok(generate_blob_transaction(blobs, TransactionSigned::arbitrary(u)?))
-    }
-}
-
-#[cfg(any(test, feature = "arbitrary"))]
-impl proptest::arbitrary::Arbitrary for BlobTransaction {
-    type Parameters = ParamsFor<String>;
-    fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-        (
-            proptest_vec(proptest_vec(proptest_any::<u8>(), BYTES_PER_BLOB), 1..=5),
-            proptest_any::<TransactionSigned>(),
-        )
-            .prop_map(move |(blobs, tx)| {
-                let blobs = blobs
-                    .into_iter()
-                    .map(|mut blob| {
-                        let mut arr = [0u8; BYTES_PER_BLOB];
-
-                        // Ensure that the blob is canonical by ensuring that
-                        // each field element contained in the blob is < BLS_MODULUS
-                        for i in 0..(FIELD_ELEMENTS_PER_BLOB as usize) {
-                            blob[i * BYTES_PER_FIELD_ELEMENT] = 0;
-                        }
-
-                        arr.copy_from_slice(blob.as_slice());
-                        arr.into()
-                    })
-                    .collect();
-
-                generate_blob_transaction(blobs, tx)
-            })
-            .boxed()
-    }
-
-    type Strategy = BoxedStrategy<BlobTransaction>;
-}
-
-#[cfg(any(test, feature = "arbitrary"))]
-fn generate_blob_transaction(blobs: Vec<Blob>, transaction: TransactionSigned) -> BlobTransaction {
-    let kzg_settings = KZG_TRUSTED_SETUP.clone();
-
-    let commitments: Vec<Bytes48> = blobs
-        .iter()
-        .map(|blob| KzgCommitment::blob_to_kzg_commitment(blob.clone(), &kzg_settings).unwrap())
-        .map(|commitment| commitment.to_bytes())
-        .collect();
-
-    let proofs: Vec<Bytes48> = blobs
-        .iter()
-        .zip(commitments.iter())
-        .map(|(blob, commitment)| {
-            KzgProof::compute_blob_kzg_proof(blob.clone(), *commitment, &kzg_settings).unwrap()
-        })
-        .map(|proof| proof.to_bytes())
-        .collect();
-
-    let sidecar = BlobTransactionSidecar { blobs, commitments, proofs };
-
-    BlobTransaction { transaction, sidecar }
-}
 #[cfg(test)]
 mod test {
     use crate::{message::RequestPair, GetPooledTransactions, PooledTransactions};
