+++ crates/primitives/src/transaction/eip4844.rs
@@ -1,10 +1,17 @@
 use super::access_list::AccessList;
-use crate::{constants::eip4844::DATA_GAS_PER_BLOB, Bytes, ChainId, TransactionKind, H256};
-use c_kzg::{Blob, Bytes48, BYTES_PER_BLOB, BYTES_PER_COMMITMENT, BYTES_PER_PROOF};
+use crate::{
+    constants::eip4844::DATA_GAS_PER_BLOB,
+    kzg::{
+        self, Blob, Bytes48, KzgCommitment, KzgProof, KzgSettings, BYTES_PER_BLOB,
+        BYTES_PER_COMMITMENT, BYTES_PER_FIELD_ELEMENT, BYTES_PER_PROOF, FIELD_ELEMENTS_PER_BLOB,
+    },
+    kzg_to_versioned_hash, Bytes, ChainId, Signature, Transaction, TransactionKind,
+    TransactionSigned, TransactionSignedNoHash, TxType, EIP4844_TX_TYPE_ID, H256,
+};
 use reth_codecs::{main_codec, Compact};
-use reth_rlp::{Decodable, DecodeError, Encodable};
+use reth_rlp::{Decodable, DecodeError, Encodable, Header};
 use serde::{Deserialize, Serialize};
-use std::mem;
+use std::{mem, ops::Deref};
 
 /// [EIP-4844 Blob Transaction](https://eips.ethereum.org/EIPS/eip-4844#blob-transaction)
 ///
@@ -149,6 +156,281 @@ impl TxEip4844 {
         self.input.len() +  // input
         self.blob_versioned_hashes.capacity() * mem::size_of::<H256>() + // blob hashes size
         mem::size_of::<u128>() // max_fee_per_data_gas
+    }
+}
+
+/// An error that can occur when validating a [BlobTransaction].
+#[derive(Debug)]
+pub enum BlobTransactionValidationError {
+    /// An error returned by the [kzg] library
+    KZGError(kzg::Error),
+    /// The inner transaction is not a blob transaction
+    NotBlobTransaction(TxType),
+}
+
+impl From<kzg::Error> for BlobTransactionValidationError {
+    fn from(value: kzg::Error) -> Self {
+        Self::KZGError(value)
+    }
+}
+
+/// A response to [`GetPooledTransactions`] that includes blob data, their commitments, and their
+/// corresponding proofs.
+///
+/// This is defined in [EIP-4844](https://eips.ethereum.org/EIPS/eip-4844#networking) as an element
+/// of a [PooledTransactions] response.
+///
+/// NOTE: This contains a [TransactionSigned], which could be a non-4844 transaction type, even
+/// though that would not make sense. This type is meant to be constructed using decoding methods,
+/// which should always construct the [TransactionSigned] with an EIP-4844 transaction.
+#[derive(Clone, Debug, PartialEq, Eq, Default, Serialize, Deserialize)]
+pub struct BlobTransaction {
+    /// The transaction payload.
+    pub transaction: TransactionSigned,
+    /// The transaction's blob sidecar.
+    pub sidecar: BlobTransactionSidecar,
+}
+
+impl BlobTransaction {
+    /// Verifies that the transaction's blob data, commitments, and proofs are all valid.
+    ///
+    /// Takes as input the [KzgSettings], which should contain the the parameters derived from the
+    /// KZG trusted setup.
+    ///
+    /// This ensures that the blob transaction payload has the same number of blob data elements,
+    /// commitments, and proofs. Each blob data element is verified against its commitment and
+    /// proof.
+    ///
+    /// Returns `false` if any blob KZG proof in the response fails to verify, or if the versioned
+    /// hashes in the transaction do not match the actual commitment versioned hashes.
+    pub fn validate(
+        &self,
+        proof_settings: &KzgSettings,
+    ) -> Result<bool, BlobTransactionValidationError> {
+        let inner_tx = match &self.transaction.transaction {
+            Transaction::Eip4844(blob_tx) => blob_tx,
+            non_blob_tx => {
+                return Err(BlobTransactionValidationError::NotBlobTransaction(
+                    non_blob_tx.tx_type(),
+                ))
+            }
+        };
+
+        // Ensure the versioned hashes and commitments have the same length
+        if inner_tx.blob_versioned_hashes.len() != self.sidecar.commitments.len() {
+            return Err(kzg::Error::MismatchLength(format!(
+                "There are {} versioned commitment hashes and {} commitments",
+                inner_tx.blob_versioned_hashes.len(),
+                self.sidecar.commitments.len()
+            ))
+            .into())
+        }
+
+        // zip and iterate, calculating versioned hashes
+        for (versioned_hash, commitment) in
+            inner_tx.blob_versioned_hashes.iter().zip(self.sidecar.commitments.iter())
+        {
+            // convert to KzgCommitment
+            let commitment = KzgCommitment::from(*commitment.deref());
+
+            // Calculate the versioned hash
+            //
+            // TODO: should this method distinguish the type of validation failure? For example
+            // whether a certain versioned hash does not match, or whether the blob proof
+            // validation failed?
+            let calculated_versioned_hash = kzg_to_versioned_hash(commitment);
+            if *versioned_hash != calculated_versioned_hash {
+                return Ok(false)
+            }
+        }
+
+        // Verify as a batch
+        KzgProof::verify_blob_kzg_proof_batch(
+            self.sidecar.blobs.as_slice(),
+            self.sidecar.commitments.as_slice(),
+            self.sidecar.proofs.as_slice(),
+            proof_settings,
+        )
+        .map_err(Into::into)
+    }
+
+    /// Splits the [BlobTransaction] into its [TransactionSigned] and [BlobTransactionSidecar]
+    /// components.
+    pub fn into_parts(self) -> (TransactionSigned, BlobTransactionSidecar) {
+        (self.transaction, self.sidecar)
+    }
+
+    /// Encodes the [BlobTransaction] fields as RLP, with a tx type. If `with_header` is `false`,
+    /// the following will be encoded:
+    /// `tx_type (0x03) || rlp([transaction_payload_body, blobs, commitments, proofs])`
+    ///
+    /// If `with_header` is `true`, the following will be encoded:
+    /// `rlp(tx_type (0x03) || rlp([transaction_payload_body, blobs, commitments, proofs]))`
+    ///
+    /// NOTE: The header will be a byte string header, not a list header.
+    pub fn encode_with_type_inner(&self, out: &mut dyn bytes::BufMut, with_header: bool) {
+        // Calculate the length of:
+        // `tx_type || rlp([transaction_payload_body, blobs, commitments, proofs])`
+        //
+        // to construct and encode the string header
+        if with_header {
+            Header {
+                list: false,
+                // add one for the tx type
+                payload_length: 1 + self.payload_len(),
+            }
+            .encode(out);
+        }
+
+        out.put_u8(EIP4844_TX_TYPE_ID);
+
+        // Now we encode the inner blob transaction:
+        self.encode_inner(out);
+    }
+
+    /// Encodes the [BlobTransaction] fields as RLP, with the following format:
+    /// `rlp([transaction_payload_body, blobs, commitments, proofs])`
+    ///
+    /// where `transaction_payload_body` is a list:
+    /// `[chain_id, nonce, max_priority_fee_per_gas, ..., y_parity, r, s]`
+    ///
+    /// Note: this should be used only when implementing other RLP encoding methods, and does not
+    /// represent the full RLP encoding of the blob transaction.
+    pub fn encode_inner(&self, out: &mut dyn bytes::BufMut) {
+        // First we construct both required list headers.
+        //
+        // The `transaction_payload_body` length is the length of the fields, plus the length of
+        // its list header.
+        let tx_header = Header {
+            list: true,
+            payload_length: self.transaction.fields_len() +
+                self.transaction.signature.payload_len(),
+        };
+
+        let tx_length = tx_header.length() + tx_header.payload_length;
+
+        // The payload length is the length of the `tranascation_payload_body` list, plus the
+        // length of the blobs, commitments, and proofs.
+        let payload_length = tx_length + self.sidecar.fields_len();
+
+        // First we use the payload len to construct the first list header
+        let blob_tx_header = Header { list: true, payload_length };
+
+        // Encode the blob tx header first
+        blob_tx_header.encode(out);
+
+        // Encode the inner tx list header, then its fields
+        tx_header.encode(out);
+        self.transaction.encode_fields(out);
+
+        // Encode the blobs, commitments, and proofs
+        self.sidecar.encode_inner(out);
+    }
+
+    /// Ouputs the length of the RLP encoding of the blob transaction, including the tx type byte,
+    /// optionally including the length of a wrapping string header. If `with_header` is `false`,
+    /// the length of the following will be calculated:
+    /// `tx_type (0x03) || rlp([transaction_payload_body, blobs, commitments, proofs])`
+    ///
+    /// If `with_header` is `true`, the length of the following will be calculated:
+    /// `rlp(tx_type (0x03) || rlp([transaction_payload_body, blobs, commitments, proofs]))`
+    pub fn payload_len_with_type(&self, with_header: bool) -> usize {
+        if with_header {
+            // Construct a header and use that to calculate the total length
+            let wrapped_header = Header {
+                list: false,
+                // add one for the tx type byte
+                payload_length: 1 + self.payload_len(),
+            };
+
+            // The total length is now the length of the header plus the length of the payload
+            // (which includes the tx type byte)
+            wrapped_header.length() + wrapped_header.payload_length
+        } else {
+            // Just add the length of the tx type to the payload length
+            1 + self.payload_len()
+        }
+    }
+
+    /// Outputs the length of the RLP encoding of the blob transaction with the following format:
+    /// `rlp([transaction_payload_body, blobs, commitments, proofs])`
+    ///
+    /// where `transaction_payload_body` is a list:
+    /// `[chain_id, nonce, max_priority_fee_per_gas, ..., y_parity, r, s]`
+    ///
+    /// Note: this should be used only when implementing other RLP encoding length methods, and
+    /// does not represent the full RLP encoding of the blob transaction.
+    pub fn payload_len(&self) -> usize {
+        // The `transaction_payload_body` length is the length of the fields, plus the length of
+        // its list header.
+        let tx_header = Header {
+            list: true,
+            payload_length: self.transaction.fields_len() +
+                self.transaction.signature.payload_len(),
+        };
+
+        let tx_length = tx_header.length() + tx_header.payload_length;
+
+        // The payload length is the length of the `tranascation_payload_body` list, plus the
+        // length of the blobs, commitments, and proofs.
+        tx_length + self.sidecar.fields_len()
+    }
+
+    /// Decodes a [BlobTransaction] from RLP. This expects the encoding to be:
+    /// `rlp([transaction_payload_body, blobs, commitments, proofs])`
+    ///
+    /// where `transaction_payload_body` is a list:
+    /// `[chain_id, nonce, max_priority_fee_per_gas, ..., y_parity, r, s]`
+    ///
+    /// Note: this should be used only when implementing other RLP decoding methods, and does not
+    /// represent the full RLP decoding of the [PooledTransactionResponse] type.
+    pub fn decode_inner(data: &mut &[u8]) -> Result<Self, DecodeError> {
+        // decode the _first_ list header for the rest of the transaction
+        let header = Header::decode(data)?;
+        if !header.list {
+            return Err(DecodeError::Custom("PooledTransactions blob tx must be encoded as a list"))
+        }
+
+        // Now we need to decode the inner 4844 transaction and its signature:
+        //
+        // `[chain_id, nonce, max_priority_fee_per_gas, ..., y_parity, r, s]`
+        let header = Header::decode(data)?;
+        if !header.list {
+            return Err(DecodeError::Custom(
+                "PooledTransactions inner blob tx must be encoded as a list",
+            ))
+        }
+
+        // inner transaction
+        let transaction = Transaction::Eip4844(TxEip4844::decode_inner(data)?);
+
+        // signature
+        let signature = Signature::decode(data)?;
+
+        // construct the tx now that we've decoded the fields in order
+        let tx_no_hash = TransactionSignedNoHash { transaction, signature };
+
+        // All that's left are the blobs, commitments, and proofs
+        let sidecar = BlobTransactionSidecar::decode_inner(data)?;
+
+        // # Calculating the hash
+        //
+        // The full encoding of the [PooledTransaction] response is:
+        // `tx_type (0x03) || rlp([tx_payload_body, blobs, commitments, proofs])`
+        //
+        // The transaction hash however, is:
+        // `keccak256(tx_type (0x03) || rlp(tx_payload_body))`
+        //
+        // Note that this is `tx_payload_body`, not `[tx_payload_body]`, which would be
+        // `[[chain_id, nonce, max_priority_fee_per_gas, ...]]`, i.e. a list within a list.
+        //
+        // Because the pooled transaction encoding is different than the hash encoding for
+        // EIP-4844 transactions, we do not use the original buffer to calculate the hash.
+        //
+        // Instead, we use [TransactionSignedNoHash] which will encode the transaction internally.
+        let signed_tx = tx_no_hash.with_hash();
+
+        Ok(Self { transaction: signed_tx, sidecar })
     }
 }
